all: Add CN_779_787 band defition
