+++ pkg/band/all.go
@@ -43,6 +43,15 @@ var All = map[string]map[ttnpb.PHYVersion]Band{
 		ttnpb.RP001_V1_1_REV_A:   CN_470_510_RP1_v1_1_RevA,
 		ttnpb.RP001_V1_1_REV_B:   CN_470_510_RP1_v1_1_RevB,
 	},
+	CN_779_787: {
+		ttnpb.TS001_V1_0:         CN_779_787_RP1_V1_0,
+		ttnpb.TS001_V1_0_1:       CN_779_787_RP1_V1_0_1,
+		ttnpb.RP001_V1_0_2:       CN_779_787_RP1_V1_0_2,
+		ttnpb.RP001_V1_0_2_REV_B: CN_779_787_RP1_V1_0_2_RevB,
+		ttnpb.RP001_V1_0_3_REV_A: CN_779_787_RP1_V1_0_3_RevA,
+		ttnpb.RP001_V1_1_REV_A:   CN_779_787_RP1_V1_1_RevA,
+		ttnpb.RP001_V1_1_REV_B:   CN_779_787_RP1_V1_1_RevB,
+	},
 }
 
 // Get returns the band if it was found, and returns an error otherwise.
